done Task1 (mistake - 1 in the end)

+++ src/Pirogova/Lesson4/Task1.java
@@ -3,8 +3,8 @@ package Pirogova.Lesson4;
 public class Task1 {
     public static void main(String[] args) {
         int d = 1; //количество дней
-        int l = 10; //расстояние в первый день
-        int s = 0; //суммарное расстояние
+        double l = 10; //расстояние в первый день
+        double s = 0; //суммарное расстояние
         do {
             s+=l;
             l*=1.1;
